collapse ChatHandlerProvider interface
