+++ services/src/main/java/org/apache/druid/cli/CliPeon.java
@@ -113,8 +113,6 @@ import org.apache.druid.segment.loading.OmniDataSegmentKiller;
 import org.apache.druid.segment.loading.OmniDataSegmentMover;
 import org.apache.druid.segment.loading.StorageLocation;
 import org.apache.druid.segment.realtime.ChatHandlerProvider;
-import org.apache.druid.segment.realtime.LocalChatHandlerProvider;
-import org.apache.druid.segment.realtime.NoopChatHandlerProvider;
 import org.apache.druid.segment.realtime.appenderator.AppenderatorsManager;
 import org.apache.druid.segment.realtime.appenderator.PeonAppenderatorsManager;
 import org.apache.druid.server.DruidNode;
@@ -419,24 +417,7 @@ public class CliPeon extends GuiceRunnable
 
   static void bindChatHandler(Binder binder)
   {
-    PolyBind.createChoice(
-        binder,
-        "druid.indexer.task.chathandler.type",
-        Key.get(ChatHandlerProvider.class),
-        Key.get(LocalChatHandlerProvider.class)
-    );
-    final MapBinder<String, ChatHandlerProvider> handlerProviderBinder =
-        PolyBind.optionBinder(binder, Key.get(ChatHandlerProvider.class));
-    handlerProviderBinder
-        .addBinding("local")
-        .to(LocalChatHandlerProvider.class)
-        .in(LazySingleton.class);
-    handlerProviderBinder
-        .addBinding("noop")
-        .to(NoopChatHandlerProvider.class)
-        .in(LazySingleton.class);
-    binder.bind(LocalChatHandlerProvider.class).in(LazySingleton.class);
-    binder.bind(NoopChatHandlerProvider.class).in(LazySingleton.class);
+    binder.bind(ChatHandlerProvider.class).in(LazySingleton.class);
   }
 
   static void bindPeonDataSegmentHandlers(Binder binder)
